Creating website for portfolio, using the "MEAN"
The website you are currently browsing and its content have been designed for the purpose of learning the "MEAN" stack architecture. The front-end utilizes "Angular," while the back-end is implemented with Node.js using the "Express" module, and "MongoDB" has been chosen as the database. This website, which does not require an intensive back-end workload, includes an admin panel for adding, editing, or deleting information for all cards under the "Apps" section. Within a few seconds, a new card can be added and its interactions with other cards can be configured through this panel.